Refactor Blizz controllers to base controller, add secrets access
