+++ wow-dashboard/Controllers/JournalInstancesController.cs
@@ -2,25 +2,20 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Net.Http;
-using System.Net.Http.Headers;
-using System.Text;
 using System.Text.Json;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Configuration;
 using wow_dashboard.Models.BlizzardData;
 
 namespace wow_dashboard.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
-    public class JournalInstancesController : ControllerBase
+    public class JournalInstancesController : BaseBlizzardApiController
     {
-        private readonly IHttpClientFactory _clientFactory;
-
-        public JournalInstancesController(IHttpClientFactory clientFactory)
-        {
-            _clientFactory = clientFactory;
-        }
+        public JournalInstancesController(IHttpClientFactory clientFactory, IConfiguration configuration)
+            : base(clientFactory, configuration) { }
 
         [HttpGet("{id}")]
         public async Task<ActionResult<JournalInstance>> GetJournalInstance(int id)
@@ -92,45 +87,5 @@ namespace wow_dashboard.Controllers
                 throw new Exception("There was an error with your request: " + ex.Message);
             }
         }
-
-        internal async Task<string> GetAccessTokenAsync()
-        {
-            var client = _clientFactory.CreateClient();
-
-            var clientId = "";
-            var clientSecret = "";
-
-            using var request = new HttpRequestMessage(new HttpMethod("POST"), "https://us.battle.net/oauth/token");
-
-            var base64authorization = Convert.ToBase64String(Encoding.ASCII.GetBytes($"{clientId}:{clientSecret}"));
-            
-            request.Headers.TryAddWithoutValidation("Authorization", $"Basic {base64authorization}");
-
-            request.Content = new StringContent("grant_type=client_credentials");
-            request.Content.Headers.ContentType = MediaTypeHeaderValue.Parse("application/x-www-form-urlencoded");
-
-            try
-            {
-                var response = await client.SendAsync(request);
-
-                if (response != null)
-                {
-                    var content = response.Content.ReadAsStreamAsync();
-                    var tokenResponse = await JsonSerializer.DeserializeAsync<AccessTokenResponse>(await content);
-
-                    return tokenResponse.Token;
-                }
-                else
-                {
-                    throw new Exception("The response was empty.");
-                }
-            }
-            catch (Exception ex)
-            {
-                throw new Exception("There was an error with your request: " + ex.Message);
-            }
-
-        }
-
     }
 }
